pyproject.toml
[project]
name = "tourguide"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.2",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "pgvector>=0.4.2",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=5.2.0",
    "numpy",
    "scipy",
    "transformers",
    "torch",
    "uvicorn[standard]>=0.38.0",
]
